organize chart func
