+++ packages/client/hmi-client/src/services/models/funman-service.ts
@@ -2,7 +2,6 @@ import { logger } from '@/utils/logger';
 import API from '@/api/api';
 import type { FunmanPostQueriesRequest } from '@/types/Types';
 import * as d3 from 'd3';
-import { Dictionary, groupBy } from 'lodash';
 import { ConstraintGroup, ConstraintType } from '@/components/workflow/ops/funman/funman-operation';
 
 // Partially typing Funman response
@@ -18,7 +17,7 @@ export interface FunmanBox {
 	schedule: any;
 	points: any;
 }
-export interface FunmanProcessedData {
+export interface ProcessedFunmanResult {
 	boxes: FunmanBox[];
 	points: any[];
 	states: string[];
@@ -138,7 +137,7 @@ export const processFunman = (result: any) => {
 		});
 	});
 
-	return { boxes, points, states: stateIds, trajs } as FunmanProcessedData;
+	return { boxes, points, states: stateIds, trajs } as ProcessedFunmanResult;
 };
 
 interface FunmanBoundingBox {
@@ -148,7 +147,7 @@ interface FunmanBoundingBox {
 	x2: number;
 	y2: number;
 }
-export const getBoxes = (processedData: FunmanProcessedData, param1: string, param2: string, boxType: string) => {
+export const getBoxes = (processedData: ProcessedFunmanResult, param1: string, param2: string, boxType: string) => {
 	const result: FunmanBoundingBox[] = [];
 
 	const temp = processedData.boxes
@@ -175,236 +174,6 @@ export const getBoxes = (processedData: FunmanProcessedData, param1: string, par
 	return result;
 };
 
-enum BoxType {
-	Satisfactory = 'Satisfactory',
-	Unsatisfactory = 'Unsatisfactory',
-	Ambiguous = 'Ambiguous',
-	ModelChecks = 'Model checks'
-}
-
-export function createFunmanStateChart(
-	data: FunmanProcessedData,
-	constraints: ConstraintGroup[],
-	stateId: string,
-	boxId: string
-) {
-	console.log(data, boxId);
-
-	const boxLines = data.trajs.map((traj) => {
-		const boxType = traj.label === 'true' ? BoxType.Satisfactory : BoxType.Unsatisfactory;
-		return { timepoints: traj.timestep, value: traj[stateId], boxType };
-	});
-	// Find min/max values to set an appropriate viewing range for y-axis
-	const minYValue = Math.floor(Math.min(...boxLines.map((d) => d.value)));
-	const maxYValue = Math.ceil(Math.max(...boxLines.map((d) => d.value)));
-
-	console.log(boxLines, minYValue, maxYValue);
-
-	// Show checks for the selected state
-	const stateIdConstraints = constraints.filter((c: ConstraintGroup) => c.variables.includes(stateId));
-	const modelChecks = stateIdConstraints.map((c: any) => ({
-		startX: c.timepoints.lb,
-		endX: c.timepoints.ub,
-		startY: c.additive_bounds.lb,
-		endY: c.additive_bounds.ub,
-		boxType: BoxType.ModelChecks
-	}));
-
-	return {
-		width: 600,
-		height: 300,
-		layer: [
-			{
-				mark: 'rect',
-				data: { values: modelChecks },
-				encoding: {
-					x: { field: 'startX', type: 'quantitative' },
-					x2: { field: 'endX', type: 'quantitative' },
-					y: { field: 'startY', type: 'quantitative' },
-					y2: { field: 'endY', type: 'quantitative' }
-				}
-			},
-			{
-				mark: 'line',
-				data: { values: boxLines },
-				encoding: {
-					x: { field: 'timepoints', type: 'quantitative' },
-					y: { field: 'value', type: 'quantitative' }
-				}
-			}
-		],
-		encoding: {
-			x: { title: 'Timepoints' },
-			y: {
-				title: `${stateId} (persons)`,
-				scale: { domain: [minYValue, maxYValue] }
-			},
-			color: {
-				field: 'boxType',
-				legend: { orient: 'top', direction: 'horizontal', title: null },
-				scale: {
-					domain: [BoxType.Satisfactory, BoxType.Unsatisfactory, BoxType.Ambiguous, BoxType.ModelChecks],
-					range: ['#1B8073', '#FFAB00', '#CCC569', '#A4CEFF54'] // Specify colors for each boxType
-				}
-			}
-		}
-	};
-}
-
-export const renderFumanTrajectories = (
-	element: HTMLElement,
-	processedData: FunmanProcessedData,
-	state: string,
-	selectedBoxId: string,
-	options: RenderOptions
-) => {
-	const width = options.width;
-	const height = options.height;
-	const topMargin = 10;
-	const rightMargin = 30;
-	const leftMargin = 35;
-	const bottomMargin = 30;
-	const { trajs, states } = processedData;
-
-	const elemSelection = d3.select(element);
-	d3.select(element).selectAll('*').remove();
-	const svg = elemSelection.append('svg').attr('width', width).attr('height', height);
-
-	const points: Dictionary<any> = groupBy(trajs, 'boxId');
-
-	// Find max/min across timesteps
-	const xDomain = d3.extent(trajs.map((d) => d.timestep)) as [number, number];
-
-	// Find max/min across all state values
-	const yDomain = d3.extent(trajs.map((d) => states.filter((s) => s === state).map((s) => d[s])).flat()) as [
-		number,
-		number
-	];
-
-	const xScale = d3
-		.scaleLinear()
-		.domain(xDomain)
-		.range([leftMargin, width - rightMargin]);
-
-	const yScale = d3
-		.scaleLinear()
-		.domain(yDomain)
-		.range([height - bottomMargin, topMargin]);
-
-	// Add the x-axis.
-	svg
-		.append('g')
-		.attr('transform', `translate(0,${height - bottomMargin})`)
-		.call(
-			d3
-				.axisBottom(xScale)
-				.ticks(width / 60)
-				.tickSizeOuter(0)
-		);
-
-	// Add the y-axis
-	svg
-		.append('g')
-		.attr('transform', `translate(${leftMargin},0)`)
-		.call(
-			d3
-				.axisLeft(yScale)
-				.ticks(height / 60)
-				.tickSizeOuter(0)
-		);
-
-	// Add label for x-axis
-	svg
-		.append('text')
-		.attr('class', 'x label')
-		.attr('text-anchor', 'end')
-		.attr('x', width / 2)
-		.attr('y', height - 2)
-		.text('Timestep');
-
-	const pathFn = d3
-		.line<{ x: number; y: number }>()
-		.x((d) => xScale(d.x))
-		.y((d) => yScale(d.y));
-
-	Object.keys(points).forEach((boxId) => {
-		const label = points[boxId][0].label;
-		let path = points[boxId].map((p: any) => ({ x: p.timestep, y: p[state] }));
-
-		// FIXME: funman can set the value to 0 at time t, if the trajectory ends at t
-		// This makes the linecharts really weird, until this is addressed we will ignore
-		// the last point if it is 0
-		const n = points[boxId][0].n;
-		if (n >= 0) {
-			path = path.filter((_d: any, i: number) => i <= n);
-		}
-
-		if (path.length > 1) {
-			svg
-				.append('g')
-				.append('path')
-				.attr('d', pathFn(path))
-				.style('stroke', label === 'true' ? 'teal' : 'orange')
-				.style('stroke-width', () => {
-					if (selectedBoxId === '' || selectedBoxId === boxId) return 2.0;
-					return 1.0;
-				})
-				.style('opacity', () => {
-					if (selectedBoxId === '' || selectedBoxId === boxId) return 0.75;
-					return 0.05;
-				})
-				.style('fill', 'none');
-
-			svg
-				.selectAll(`.${boxId}`)
-				.data(path)
-				.enter()
-				.append('circle')
-				.attr('cx', (d: any) => xScale(d.x))
-				.attr('cy', (d: any) => yScale(d.y))
-				.attr('r', 2)
-				.style('opacity', () => {
-					if (selectedBoxId === '' || selectedBoxId === boxId) return 0.75;
-					return 0.05;
-				})
-				.style('fill', label === 'true' ? 'teal' : 'orange');
-		} else if (path.length === 1) {
-			svg
-				.append('g')
-				.append('circle')
-				.attr('cx', xScale(path[0].x))
-				.attr('cy', yScale(path[0].y))
-				.attr('r', 4)
-				.style('fill', '#888');
-		}
-	});
-
-	// Render constraints
-	// Since this is variable-vs-time, we can't display constraints that involve more
-	// than one variable, and just the selected variable
-	if (options.constraints && options.constraints.length > 0) {
-		const singleVariableConstraints = options.constraints.filter((d) => d.variable).filter((d) => d.variable === state);
-
-		svg
-			.selectAll('.constraint-box')
-			.data(singleVariableConstraints)
-			.enter()
-			.append('rect')
-			.classed('constraint-box', true)
-			.attr('x', (d) => xScale(d.timepoints?.lb as number))
-			.attr('y', (d) => yScale(d.interval?.ub as number))
-			.attr('width', (d) => {
-				const w = xScale(d.timepoints?.ub as number) - xScale(d.timepoints?.lb as number);
-				return w;
-			})
-			.attr('height', (d) => Math.abs(yScale(d.interval?.ub as number) - yScale(d.interval?.lb as number)))
-			.style('fill', 'teal')
-			.style('fill-opacity', 0.3);
-	}
-
-	return svg;
-};
-
 const getBoxesDomain = (boxes: FunmanBoundingBox[]) => {
 	let minX = Number.MAX_VALUE;
 	let maxX = Number.MIN_VALUE;
@@ -423,7 +192,7 @@ const getBoxesDomain = (boxes: FunmanBoundingBox[]) => {
 
 export const renderFunmanBoundaryChart = (
 	element: HTMLElement,
-	processedData: FunmanProcessedData,
+	processedData: ProcessedFunmanResult,
 	param1: string,
 	param2: string,
 	selectedBoxId: string,
